refactor: split record form components
